requirements.txt
bleak==0.10.0
service-identity==18.1.0
paho-mqtt==1.5.1
